# Small performance improvements to some schedulers