--- a/src/diffusers/schedulers/scheduling_dpmsolver_multistep.py
+++ b/src/diffusers/schedulers/scheduling_dpmsolver_multistep.py
@@ -310,30 +310,34 @@ class DPMSolverMultistepScheduler(SchedulerMixin, ConfigMixin):
         if self.config.algorithm_type == "dpmsolver++":
             # See https://arxiv.org/abs/2211.01095 for detailed derivations
             if self.config.solver_type == "midpoint":
+                alpha_t_exp_h = (alpha_t * (torch.exp(-h) - 1.0))
                 x_t = (
                     (sigma_t / sigma_s0) * sample
-                    - (alpha_t * (torch.exp(-h) - 1.0)) * D0
-                    - 0.5 * (alpha_t * (torch.exp(-h) - 1.0)) * D1
+                    - alpha_t_exp_h * D0
+                    - 0.5 * alpha_t_exp_h * D1
                 )
             elif self.config.solver_type == "heun":
+                exp_h = torch.exp(-h) - 1.0
                 x_t = (
                     (sigma_t / sigma_s0) * sample
-                    - (alpha_t * (torch.exp(-h) - 1.0)) * D0
-                    + (alpha_t * ((torch.exp(-h) - 1.0) / h + 1.0)) * D1
+                    - (alpha_t * exp_h) * D0
+                    + (alpha_t * (exp_h / h + 1.0)) * D1
                 )
         elif self.config.algorithm_type == "dpmsolver":
             # See https://arxiv.org/abs/2206.00927 for detailed derivations
             if self.config.solver_type == "midpoint":
+                sigma_t_exp_h = (sigma_t * (torch.exp(h) - 1.0))
                 x_t = (
                     (alpha_t / alpha_s0) * sample
-                    - (sigma_t * (torch.exp(h) - 1.0)) * D0
-                    - 0.5 * (sigma_t * (torch.exp(h) - 1.0)) * D1
+                    - sigma_t_exp_h * D0
+                    - 0.5 * sigma_t_exp_h * D1
                 )
             elif self.config.solver_type == "heun":
+                exp_h = (torch.exp(h) - 1.0)
                 x_t = (
                     (alpha_t / alpha_s0) * sample
-                    - (sigma_t * (torch.exp(h) - 1.0)) * D0
-                    - (sigma_t * ((torch.exp(h) - 1.0) / h - 1.0)) * D1
+                    - (sigma_t * exp_h) * D0
+                    - (sigma_t * (exp_h / h - 1.0)) * D1
                 )
         return x_t
 
@@ -376,19 +380,21 @@ class DPMSolverMultistepScheduler(SchedulerMixin, ConfigMixin):
         D2 = (1.0 / (r0 + r1)) * (D1_0 - D1_1)
         if self.config.algorithm_type == "dpmsolver++":
             # See https://arxiv.org/abs/2206.00927 for detailed derivations
+            exp_h = torch.exp(-h) - 1.0
             x_t = (
                 (sigma_t / sigma_s0) * sample
-                - (alpha_t * (torch.exp(-h) - 1.0)) * D0
-                + (alpha_t * ((torch.exp(-h) - 1.0) / h + 1.0)) * D1
-                - (alpha_t * ((torch.exp(-h) - 1.0 + h) / h**2 - 0.5)) * D2
+                - (alpha_t * exp_h) * D0
+                + (alpha_t * (exp_h / h + 1.0)) * D1
+                - (alpha_t * ((exp_h + h) / h**2 - 0.5)) * D2
             )
         elif self.config.algorithm_type == "dpmsolver":
             # See https://arxiv.org/abs/2206.00927 for detailed derivations
+            exp_h = torch.exp(h) - 1.0
             x_t = (
                 (alpha_t / alpha_s0) * sample
-                - (sigma_t * (torch.exp(h) - 1.0)) * D0
-                - (sigma_t * ((torch.exp(h) - 1.0) / h - 1.0)) * D1
-                - (sigma_t * ((torch.exp(h) - 1.0 - h) / h**2 - 0.5)) * D2
+                - (sigma_t * exp_h) * D0
+                - (sigma_t * (exp_h / h - 1.0)) * D1
+                - (sigma_t * ((exp_h - h) / h**2 - 0.5)) * D2
             )
         return x_t