@HuggingFaceDocBuilderDev 2022-11-16
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/diffusers/pr_1303). All of your documentation changes will be reflected on that endpoint.