@patrickvonplaten 2022-11-16
Uff a bit worried here. Not that small changes is the sequence of operations can lead to large changes in the output . The very high precision tests are only run for the slow tests. 

Also not 100% sure how important this is as the majority of the time will always be taken by the model's forward pass no?

I'm sorry, but I'm not really in favor here :-/